+++ inst/shiny/modules/tab_explore.R
@@ -189,15 +189,8 @@ tab_explore_server <- function(id, pknca_data, extra_group_vars) {
     # Save the objects for the ZIP folder whenever they change
     observe({
       req(individualplot(), meanplot())
-
-      # Individual plots
       session$userData$results$exploration$individualplot <- individualplot()
-      session$userData$results$exploration$individualplot_plotly <- ggplotly(individualplot(),
-                                                                             tooltip = "text")
-
-      # Mean plots
       session$userData$results$exploration$meanplot <- meanplot()
-      session$userData$results$exploration$meanplot_plotly <- ggplotly(meanplot(), tooltip = "text")
     })
 
     # Render the mean plot output in plotly