+++ inst/shiny/modules/tab_nca.R
@@ -63,7 +63,7 @@ tab_nca_ui <- function(id) {
   )
 }
 
-tab_nca_server <- function(id, pknca_data, extra_group_vars) {
+tab_nca_server <- function(id, pknca_data, extra_group_vars, settings_override) {
   moduleServer(id, function(input, output, session) {
     ns <- session$ns
 
@@ -75,20 +75,32 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
     adnca_data <- reactive(pknca_data()$conc$data)
 
     # #' NCA Setup module
-    nca_setup <- setup_server("nca_setup", adnca_data, pknca_data)
+    nca_setup <- setup_server(
+      "nca_setup",
+      adnca_data,
+      pknca_data,
+      extra_group_vars,
+      settings_override
+    )
 
     processed_pknca_data <- nca_setup$processed_pknca_data
     settings <- nca_setup$settings
 
     ratio_table <- nca_setup$ratio_table
     slope_rules <- nca_setup$slope_rules
+
     session$userData$settings <- list(
       settings =  settings,
-      slope_rules = slope_rules$manual_slopes
+      slope_rules = slope_rules
     ) # This will be saved in the results zip folder
 
+    # This will be saved in the results zip folder
+    session$userData$settings <- settings
+    session$userData$ratio_table <- ratio_table
+    session$userData$slope_rules <- slope_rules
+
     reactable_server("manual_slopes", slope_rules$manual_slopes,
-                     columns = NULL)
+                    columns = NULL)
 
     # List all irrelevant warnings to suppres in the NCA calculation
     irrelevant_regex_warnings <- c(
@@ -129,13 +141,15 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
         res <- withCallingHandlers({
           processed_pknca_data %>%
             filter_slopes(
-              slope_rules$manual_slopes(),
-              slope_rules$profiles_per_subject(),
-              slope_rules$slopes_groups(),
+              slope_rules()$manual_slopes,
+              slope_rules()$profiles_per_subject,
+              slope_rules()$slopes_groups,
               check_reasons = TRUE
             ) %>%
             # Perform PKNCA parameter calculations
-            PKNCA_calculate_nca() %>%
+            PKNCA_calculate_nca(
+              blq_rule = settings()$data_imputation$blq_imputation_rule
+            ) %>%
             # Add bioavailability results if requested
             add_f_to_pknca_results(settings()$bioavailability) %>%
             # Apply standard CDISC names
@@ -149,7 +163,9 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
                 purrr::map(\(x) x$threshold)
             ) %>%
             # Add parameter ratio calculations
-            calculate_table_ratios_app(ratio_table = ratio_table())
+            calculate_table_ratios(ratio_table = ratio_table()) %>%
+            # Keep only parameters requested by the user
+            remove_pp_not_requested()
         },
         warning = function(w) {
           if (!grepl(paste(irrelevant_regex_warnings, collapse = "|"),
@@ -170,33 +186,6 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
 
         log_success("NCA results calculated.")
 
-        params_requested <- c(setdiff(names(PKNCA::get.interval.cols()), c("start", "end")),
-                              settings()$bioavailability)
-        # Reshape intervals, filter
-        params_not_requested <- res$data$intervals %>%
-          # add bioavailability if requested
-          mutate(!!!rlang::set_names(TRUE, settings()$bioavailability)) %>%
-          # pivot for requested params
-          pivot_longer(
-            cols = (any_of(params_requested)),
-            names_to = "PPTESTCD",
-            values_to = "is_requested"
-          ) %>%
-          # Translate terms
-          mutate(PPTESTCD = translate_terms(PPTESTCD, "PKNCA", "PPTESTCD")) %>%
-          # Group by all identifying cols EXCEPT the impute column and the value column
-          group_by(across(c(-impute, -is_requested))) %>%
-          # If all are FALSE, any(is_requested) will be FALSE.
-          summarise(
-            is_requested = any(is_requested),
-            .groups = "drop"
-          ) %>%
-          filter(!is_requested)
-
-        # Filter for requested params based on intervals
-        res$result <- res$result %>%
-          anti_join(params_not_requested, by = intersect(names(.), names(params_not_requested)))
-
         res
       }, error = function(e) {
         log_error("Error calculating NCA results:\n{conditionMessage(e)}")
@@ -208,6 +197,11 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
       })
     }) %>%
       bindEvent(input$run_nca)
+
+    observe({
+      req(res_nca())
+      session$userData$final_units <- res_nca()$data$units
+    })
 
     #' Show slopes results
     pivoted_slopes <- reactive({