+++ tests/testthat/test-pivot_wider_pknca_results.R
@@ -207,4 +207,55 @@ describe("pivot_wider_pknca_results", {
 
     expect_false("PPANMETH" %in% result)
   })
+
+  it("includes the flagged column and the flag rule columns when flag_rules are provided", {
+
+    # Prepare a PKNCA results object with the exclusions to flag
+    flag_rules <- list(
+      R2ADJ = list(is.checked = TRUE, threshold = 0.99),
+      R2 = list(is.checked = TRUE, threshold = 0.99)
+    )
+    flag_rules_pknca <- flag_rules %>%
+      purrr::keep(\(x) x$is.checked) %>%
+      purrr::map(\(x) x$threshold)
+    pknca_res <- PKNCA_hl_rules_exclusion(res = pknca_res, rules = flag_rules_pknca)
+
+    # Produce the output to be tested
+    piv_result <- pivot_wider_pknca_results(pknca_res, flag_rules = flag_rules)
+
+    # Missing results produce "MISSING" flag
+    na_result <- piv_result %>%
+      filter(is.na(R2) | is.na(R2ADJ)) %>%
+
+      # TODO (Gerardo): Exclude test on SUBJ 4 because
+      # is a weird case scenario that PKNCA does not detect
+      # It is a straight line, R2 = NA, but PKNCA does not flag exclude
+      filter(USUBJID != 4)
+
+    expect_equal(unique(na_result$flagged), "MISSING")
+
+    # Invalid rules produce "FLAGGED" flag
+    flagged_result <- piv_result %>%
+      filter(!is.na(R2) & !is.na(R2ADJ)) %>%
+      filter(R2 < flag_rules$R2$threshold | R2ADJ < flag_rules$R2ADJ$threshold)
+    expect_equal(unique(flagged_result$flagged), "FLAGGED")
+
+    # Passed rules produce "ACCEPTED" flag
+    accepted_result <- piv_result %>%
+      filter(!is.na(R2) & !is.na(R2ADJ)) %>%
+      filter(R2 >= flag_rules$R2$threshold & R2ADJ >= flag_rules$R2ADJ$threshold)
+    expect_equal(unique(accepted_result$flagged), "ACCEPTED")
+  })
+
+  it("includes extra_vars_to_keep columns when provided", {
+    pknca_res <- FIXTURE_PKNCA_RES
+    extra_vars <- c("ROUTE", "DOSEA", "ATPTREF")
+    pivoted_res <- pivot_wider_pknca_results(pknca_res, extra_vars_to_keep = extra_vars)
+
+    # All requested extra_vars should be present in the output (if present in conc data)
+    present_vars <- intersect(extra_vars, names(pknca_res$data$conc$data))
+
+    # Check all extra_vars present are added in the pivoted result
+    expect_true(all(present_vars %in% names(pivoted_res)))
+  })
 })