+++ NEWS.md
@@ -4,14 +4,19 @@
 
 * Enhancements to the slides outputs including grouping by PKNCA groups, dose profile, and additional grouping variables (#791)
 * Option to include and apply NCA flag rules with reasons (NCAwXRS) as defined by ADNCA standards. Any record populated within these columns will be excluded for the NCA (#752)
+* R script exported in ZIP folder to re-run and replicate App outputs (#789)
 * Individual and Mean plots tabs now created using the same function, so the layout and plot themes are consistent across both plots (#712)
 * New flagging rule for lambda-z calculations based on r-squared, R2 (#834)
 * New Parameter Selection section in NCA tab allowing to select parameters by study type (#795)
 * The App optionally maps end of sample collection (AEFRLT) for excretion rate parameter calculations: ERTLST, ERTMAX. (#745)
 * Option to upload multiple input files, which will be bound together to form a single ADNCA data set (#821)
+* BLQ imputation rules can be applied to the NCA via `NCA Setup > Data Imputation` (#139)
 * Section `General Exclusions` allowing to perform in-App NCA exclusions (#851)
 * Mean plots have been added in the TLGs section, with BLQ handling (#555)
 * CMAX automatically selected in box plots if available (#890)
+* Allow user to select additional `grouping variables` (chosen in the mapping) for ratio calculations (#868)
+* General button at top page to save all NCA results, settings & draft slides as a ZIP file (#638)
+* Settings are now uploaded on initial opening of the app in the data tab, and applied to the next steps (#860)
 
 
 ## Bugs fixed
@@ -26,7 +31,7 @@
 * Units table bugs fixed, so it is filtered based on NCA setup and the table is searchable for each column (#870)
 * Bug fix to allow for unrecognized units to be used in AVALU and DOSEU (#861)
 * Bug fix so NA units are allowed in the data (as per CDISC guidelines for NA samples) and not treated as a unique unit (#907)
-
+* Bug fix so app doesn't crash if NCA is rerun with an error (e.g. No exclusion REASON) (#913)
 
 # aNCA 0.1.0